requirements.txt
asgiref==3.2.7
blinker==1.4
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.41.0
django-environ==0.4.5
idna==2.9
mangopaysdk==3.8.0
oauthlib==3.1.0
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
